Custom timeout in consumeNumberMessagesFromTopics
This allows a user to specify a custom timeout in the
consumeNumberMessagesFromTopics method, instead of hard-coding 5sec.

This also adds a new resetTimeoutOnEachMessage parameter. The default
value (true) uses the old behavior, of timing out if there is a period
of silence. Setting  it to false counts the total consume time toward
the timeout. E.g. if a user calls this method with number=12 and
timeout=5.seconds and resetTimeoutOnEachMessage=false, and a producer
produces one message every 4 seconds, the method will now throw
TimeoutException after 5 seconds. The previous/default behavior would
receive all messages in 60 seconds and not throw an exception, because
there was at least one message to consume every 5 seconds.