Add a batch version of publishToKafka
It is very expensive to create a `KafkaProducer`.

For tests that send a lot of messages it is very slow
to use the original `publishToKafka` method that created
a new `KafkaProducer` for each message to send.

Now there is a batch version of `publishToKafka` that
accepts a sequence of key/value tuples and sends them
all with the same `KafkaProducer` instance.