144 Remove the coupling with ScalaTest
From what I can tell, there is no need to require that ScalaTest's
`Suite` was mixed in.  Without this, this library can be used in
more contexts without dragging in scalatest classes.

This is just food for thought.  I almost never use self-types, so I
might be missing something.  However, the following appear to be true:

1. The tests of this project pass.
2. My TestNG tests that use EmbeddedKafka#start/stop work.
3. My existing Scala project that is mixing in EmbeddedKafka to
ScalaTest tests still work.